Implement timeline-related commands

+++ src/command/CommandModifyAnimationKey.hpp
@@ -14,7 +14,7 @@ class CommandModifyAnimationKey : public BaseCommand {
 public:
     // Constructor: Replace key from cellanimIndex, animationIndex and keyIndex by newKey.
     CommandModifyAnimationKey(
-        uint16_t cellanimIndex, uint32_t animationIndex, uint32_t keyIndex,
+        uint16_t cellanimIndex, uint16_t animationIndex, uint32_t keyIndex,
         RvlCellAnim::AnimationKey newKey
     ) :
         cellanimIndex(cellanimIndex), animationIndex(animationIndex), keyIndex(keyIndex),
@@ -40,7 +40,7 @@ public:
 
 private:
     uint16_t cellanimIndex;
-    uint32_t animationIndex;
+    uint16_t animationIndex;
     uint32_t keyIndex;
 
     RvlCellAnim::AnimationKey oldKey;
